grcpPydenticSqlAlchemyAlembic/src/ex00/requirements.txt
black==24.1.1
click==8.1.7
Faker==22.6.0
grpcio==1.60.1
grpcio-tools==1.60.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
protobuf==4.25.2
python-dateutil==2.8.2
six==1.16.0
